Fix mono_marshal_alloc to accept gsize instead of gulong(unsigned long) (mono/mono#4623)
* Fix mono_marshal_alloc to accept gsize instead of platform dependent unsigned long

* fix header


Commit migrated from https://github.com/mono/mono/commit/47ca2c886d0ba4ed5cd7bf90a085c7c2f539d0f1

+++ src/mono/mono/metadata/marshal.c
@@ -10678,7 +10678,7 @@ mono_marshal_alloc_co_task_mem (size_t size)
  * mono_marshal_alloc:
  */
 void*
-mono_marshal_alloc (gulong size, MonoError *error)
+mono_marshal_alloc (gsize size, MonoError *error)
 {
 	gpointer res;
